Liberate k8s resource detection
