+++ tests/unit/common/monitored_resource_util/test_gcp_metadata_config.py
@@ -91,93 +91,6 @@ class TestGcpMetadataConfig(unittest.TestCase):
 
         self.assertDictEqual(labels_list, expected_labels)
 
-    @mock.patch.dict(
-        os.environ, {
-            'KUBERNETES_SERVICE_HOST': '127.0.0.1',
-            'CONTAINER_NAME': 'container',
-            'NAMESPACE': 'namespace',
-            'HOSTNAME': 'localhost'
-        },
-        clear=True)
-    @mock.patch('opencensus.common.monitored_resource.'
-                'gcp_metadata_config.get_request')
-    def test_get_k8s_metadata(self, http_request_mock):
-        def assign_attribute_value(*args, **kwargs):
-            attribute_uri = args[0].split('/')[-1]
-            if attribute_uri == 'id':
-                return 'my-instance'
-            elif attribute_uri == 'project-id':
-                return 'my-project'
-            elif attribute_uri == 'cluster-name':
-                return 'cluster'
-            elif attribute_uri == 'zone':
-                return 'us-east1'
-
-        http_request_mock.side_effect = assign_attribute_value
-        gcp_metadata_config.GcpMetadataConfig.inited = False
-        gcp_metadata_config.GcpMetadataConfig.is_running = False
-        gcp_metadata_config.gcp_metadata_map = {}
-
-        self.assertTrue(
-            gcp_metadata_config.GcpMetadataConfig.is_running_on_gcp())
-
-        labels_list = (gcp_metadata_config.GcpMetadataConfig()
-                       .get_k8s_metadata())
-
-        self.assertEquals(len(labels_list), 7)
-
-        expected_labels = {
-            'instance_id': 'my-instance',
-            'cluster_name': 'cluster',
-            'project_id': 'my-project',
-            'location': 'us-east1',
-            'pod_name': 'localhost',
-            'namespace_name': 'namespace',
-            'container_name': 'container'
-        }
-        self.assertDictEqual(labels_list, expected_labels)
-
-    @mock.patch.dict(
-        os.environ, {
-            'KUBERNETES_SERVICE_HOST': '127.0.0.1',
-            'NAMESPACE': 'namespace',
-            'HOSTNAME': 'localhost'
-        },
-        clear=True)
-    @mock.patch('opencensus.common.monitored_resource.'
-                'gcp_metadata_config.get_request')
-    def test_get_k8s_metadata_container_empty(self, http_request_mock):
-        def assign_attribute_value(*args, **kwargs):
-            attribute_uri = args[0].split('/')[-1]
-            if attribute_uri == 'id':
-                return 'my-instance'
-            elif attribute_uri == 'project-id':
-                return 'my-project'
-            elif attribute_uri == 'zone':
-                return 'us-east1'
-
-        http_request_mock.side_effect = assign_attribute_value
-        gcp_metadata_config.GcpMetadataConfig.inited = False
-        gcp_metadata_config.GcpMetadataConfig.is_running = False
-        gcp_metadata_config.gcp_metadata_map = {}
-
-        self.assertTrue(
-            gcp_metadata_config.GcpMetadataConfig.is_running_on_gcp())
-
-        labels_list = gcp_metadata_config.GcpMetadataConfig().get_k8s_metadata(
-        )
-
-        self.assertEquals(len(labels_list), 5)
-
-        expected_labels = {
-            'instance_id': 'my-instance',
-            'project_id': 'my-project',
-            'location': 'us-east1',
-            'pod_name': 'localhost',
-            'namespace_name': 'namespace'
-        }
-        self.assertDictEqual(labels_list, expected_labels)
-
     @mock.patch.dict(os.environ, clear=True)
     @mock.patch('opencensus.common.monitored_resource.'
                 'gcp_metadata_config.get_request')
@@ -192,5 +105,3 @@ class TestGcpMetadataConfig(unittest.TestCase):
 
         self.assertEquals(
             len(gcp_metadata_config.GcpMetadataConfig().get_gce_metadata()), 0)
-        self.assertEquals(
-            len(gcp_metadata_config.GcpMetadataConfig().get_k8s_metadata()), 0)
